Move resources to separate packages
