+++ src/main/scala/com/cognite/sdk/scala/v1/resources/ReadWritableResourceV1.scala
@@ -1,17 +1,10 @@
-package com.cognite.sdk.scala.v1
+package com.cognite.sdk.scala.v1.resources
 
-import com.cognite.sdk.scala.common.{
-  Auth,
-  CdpApiError,
-  CogniteId,
-  EitherDecoder,
-  Items,
-  ReadWritableResourceWithRetrieve
-}
-import com.softwaremill.sttp.circe._
+import com.cognite.sdk.scala.common._
 import com.softwaremill.sttp._
-import io.circe.{Decoder, Encoder}
+import com.softwaremill.sttp.circe._
 import io.circe.generic.auto._
+import io.circe.{Decoder, Encoder}
 
 abstract class ReadWritableResourceV1[R: Decoder, W: Decoder: Encoder, F[_]](
     implicit auth: Auth,
