+++ src/main/scala/com/cognite/sdk/scala/v1/resources/ResourceV1.scala
@@ -1,0 +1,8 @@
+package com.cognite.sdk.scala.v1.resources
+
+import com.cognite.sdk.scala.common.CogniteId
+
+trait ResourceV1[F[_]] {
+  def toInternalId(id: Long): CogniteId = CogniteId(id)
+  //
+}
