+++ src/main/scala/com/cognite/sdk/scala/v1/resources/assets.scala
@@ -1,6 +1,6 @@
-package com.cognite.sdk.scala.v06
+package com.cognite.sdk.scala.v1.resources
 
-import com.cognite.sdk.scala.common._
+import com.cognite.sdk.scala.common.{Auth, WithId}
 import com.softwaremill.sttp._
 import io.circe.generic.auto._
 
@@ -13,7 +13,7 @@ final case class Asset(
     description: Option[String] = None,
     metadata: Option[Map[String, String]] = None,
     source: Option[String] = None,
-    sourceId: Option[String] = None,
+    externalId: Option[String] = None,
     createdTime: Option[Long] = None,
     lastUpdatedTime: Option[Long] = None
 ) extends WithId[Long]
@@ -23,12 +23,12 @@ final case class CreateAsset(
     parentId: Option[Long] = None,
     description: Option[String] = None,
     source: Option[String] = None,
-    sourceId: Option[String] = None,
+    externalId: Option[String] = None,
     metadata: Option[Map[String, String]] = None
 )
 
 class Assets[F[_]](project: String)(implicit auth: Auth, sttpBackend: SttpBackend[F, _])
-    extends ReadWritableResourceV0_6[Asset, CreateAsset, F]
-    with ResourceV0_6[F] {
-  override val baseUri = uri"https://api.cognitedata.com/api/0.6/projects/$project/assets"
+    extends ReadWritableResourceV1[Asset, CreateAsset, F]
+    with ResourceV1[F] {
+  override val baseUri = uri"https://api.cognitedata.com/api/v1/projects/$project/assets"
 }
