+++ src/main/scala/com/cognite/sdk/scala/v1/resources/dataPoints.scala
@@ -1,48 +1,76 @@
-package com.cognite.sdk.scala.v06
+package com.cognite.sdk.scala.v1.resources
 
 import com.cognite.sdk.scala.common._
+import com.cognite.sdk.scala.v1.extractor
 import com.softwaremill.sttp._
 import com.softwaremill.sttp.circe._
 import io.circe.Decoder
 import io.circe.generic.auto._
 
-final case class DataPointsByName(
-    name: String,
+final case class DataPointsById(
+    id: Long,
     datapoints: Seq[DataPoint]
 )
 
-final case class StringDataPointsByName(
-    name: String,
+final case class DataPointsByIdResponse(
+    id: Long,
+    externalId: Option[String],
+    isString: Boolean,
+    datapoints: Seq[DataPoint]
+)
+
+final case class StringDataPointsByIdResponse(
+    id: Long,
+    externalId: Option[String],
+    isString: Boolean,
     datapoints: Seq[StringDataPoint]
 )
 
-class DataPointsResourceV0_6[F[_]](project: String)(
+final case class StringDataPointsById(
+    id: Long,
+    datapoints: Seq[StringDataPoint]
+)
+
+final case class DeleteRangeById(
+    id: Long,
+    inclusiveBegin: Long,
+    exclusiveEnd: Long
+)
+
+final case class DeleteRangeByExternalId(
+    id: String,
+    inclusiveBegin: Long,
+    exclusiveEnd: Long
+)
+
+final case class QueryRangeById(
+    id: Long,
+    start: String,
+    end: String
+)
+
+class DataPointsResourceV1[F[_]](project: String)(
     implicit auth: Auth,
     sttpBackend: SttpBackend[F, _]
-) extends Resource[F, Long, Long](auth)
-    with ResourceV0_6[F]
+) extends Resource[F, CogniteId, Long](auth)
+    with ResourceV1[F]
     with DataPointsResource[F, Long] {
-  override val baseUri = uri"https://api.cognitedata.com/api/0.6/projects/$project/timeseries"
+  override val baseUri = uri"https://api.cognitedata.com/api/v1/projects/$project/timeseries/data"
 
+  override def toInternalId(id: Long): CogniteId = CogniteId(id)
   implicit val errorOrUnitDecoder: Decoder[Either[CdpApiError[CogniteId], Unit]] =
     EitherDecoder.eitherDecoder[CdpApiError[CogniteId], Unit]
-  implicit val errorOrDataPointsByNameResponseDecoder
-      : Decoder[Either[CdpApiError[CogniteId], Data[Items[DataPointsByName]]]] =
-    EitherDecoder.eitherDecoder[CdpApiError[CogniteId], Data[Items[DataPointsByName]]]
-  implicit val errorOrStringDataPointsByNameResponseDecoder
-      : Decoder[Either[CdpApiError[CogniteId], Data[Items[StringDataPointsByName]]]] =
-    EitherDecoder.eitherDecoder[CdpApiError[CogniteId], Data[Items[StringDataPointsByName]]]
-  implicit val errorOrDataPointResponseDecoder
-      : Decoder[Either[CdpApiError[CogniteId], Data[Items[DataPoint]]]] =
-    EitherDecoder.eitherDecoder[CdpApiError[CogniteId], Data[Items[DataPoint]]]
-  implicit val errorOrStringDataPointResponseDecoder
-      : Decoder[Either[CdpApiError[CogniteId], Data[Items[StringDataPoint]]]] =
-    EitherDecoder.eitherDecoder[CdpApiError[CogniteId], Data[Items[StringDataPoint]]]
+  implicit val errorOrDataPointsByIdResponseDecoder
+      : Decoder[Either[CdpApiError[CogniteId], Items[DataPointsByIdResponse]]] =
+    EitherDecoder.eitherDecoder[CdpApiError[CogniteId], Items[DataPointsByIdResponse]]
+  implicit val errorOrStringDataPointsByIdResponseDecoder
+      : Decoder[Either[CdpApiError[CogniteId], Items[StringDataPointsByIdResponse]]] =
+    EitherDecoder.eitherDecoder[CdpApiError[CogniteId], Items[StringDataPointsByIdResponse]]
 
   def insertById(id: Long, dataPoints: Seq[DataPoint]): F[Response[Unit]] =
     request
-      .post(uri"$baseUri/$id/data")
-      .body(Items(dataPoints))
+      .post(baseUri)
+      .body(Items(Seq(DataPointsById(id, dataPoints))))
       .response(asJson[Either[CdpApiError[CogniteId], Unit]])
       .mapResponse {
         case Left(value) => throw value.error
@@ -53,8 +81,8 @@ class DataPointsResourceV0_6[F[_]](project: String)(
 
   def insertStringsById(id: Long, dataPoints: Seq[StringDataPoint]): F[Response[Unit]] =
     request
-      .post(uri"$baseUri/$id/data")
-      .body(Items(dataPoints))
+      .post(baseUri)
+      .body(Items(Seq(StringDataPointsById(id, dataPoints))))
       .response(asJson[Either[CdpApiError[CogniteId], Unit]])
       .mapResponse {
         case Left(value) => throw value.error
@@ -65,12 +93,8 @@ class DataPointsResourceV0_6[F[_]](project: String)(
 
   def deleteRangeById(id: Long, inclusiveStart: Long, exclusiveEnd: Long): F[Response[Unit]] =
     request
-      .delete(
-        uri"$baseUri/$id/data/deleterange".params(
-          ("timestampInclusiveBegin", inclusiveStart.toString),
-          ("timestampExclusiveEnd", exclusiveEnd.toString)
-        )
-      )
+      .post(uri"$baseUri/delete")
+      .body(Items(Seq(DeleteRangeById(id, inclusiveStart, exclusiveEnd))))
       .response(asJson[Either[CdpApiError[CogniteId], Unit]])
       .mapResponse {
         case Left(value) => throw value.error
@@ -81,13 +105,9 @@ class DataPointsResourceV0_6[F[_]](project: String)(
 
   def queryById(id: Long, inclusiveStart: Long, exclusiveEnd: Long): F[Response[Seq[DataPoint]]] =
     request
-      .get(
-        uri"$baseUri/$id/data".params(
-          ("start", inclusiveStart.toString),
-          ("end", exclusiveEnd.toString)
-        )
-      )
-      .response(asJson[Either[CdpApiError[CogniteId], Data[Items[DataPointsByName]]]])
+      .post(uri"$baseUri/list")
+      .body(Items(Seq(QueryRangeById(id, inclusiveStart.toString, exclusiveEnd.toString))))
+      .response(asJson[Either[CdpApiError[CogniteId], Items[DataPointsByIdResponse]]])
       .mapResponse {
         case Left(value) => throw value.error
         case Right(Left(cdpApiError)) => throw cdpApiError.asException(baseUri)
@@ -105,13 +125,9 @@ class DataPointsResourceV0_6[F[_]](project: String)(
       exclusiveEnd: Long
   ): F[Response[Seq[StringDataPoint]]] =
     request
-      .get(
-        uri"$baseUri/$id/data".params(
-          ("start", inclusiveStart.toString),
-          ("end", exclusiveEnd.toString)
-        )
-      )
-      .response(asJson[Either[CdpApiError[CogniteId], Data[Items[StringDataPointsByName]]]])
+      .post(uri"$baseUri/list")
+      .body(Items(Seq(QueryRangeById(id, inclusiveStart.toString, exclusiveEnd.toString))))
+      .response(asJson[Either[CdpApiError[CogniteId], Items[StringDataPointsByIdResponse]]])
       .mapResponse {
         case Left(value) => throw value.error
         case Right(Left(cdpApiError)) => throw cdpApiError.asException(baseUri)
@@ -126,23 +142,31 @@ class DataPointsResourceV0_6[F[_]](project: String)(
   //def deleteRangeByExternalId(start: Long, end: Long, externalId: String): F[Response[Unit]]
   def getLatestDataPointById(id: Long): F[Response[Option[DataPoint]]] =
     request
-      .get(uri"$baseUri/$id/latest")
-      .response(asJson[Either[CdpApiError[CogniteId], Data[Items[DataPoint]]]])
+      .post(uri"$baseUri/latest")
+      .body(Items(Seq(CogniteId(id))))
+      .response(asJson[Either[CdpApiError[CogniteId], Items[DataPointsByIdResponse]]])
       .mapResponse {
         case Left(value) => throw value.error
         case Right(Left(cdpApiError)) => throw cdpApiError.asException(baseUri)
-        case Right(Right(value)) => extractor.extract(value).items.headOption
+        case Right(Right(value)) =>
+          extractor.extract(value).items.headOption.flatMap(_.datapoints.headOption)
       }
       .send()
 
   def getLatestStringDataPointById(id: Long): F[Response[Option[StringDataPoint]]] =
     request
-      .get(uri"$baseUri/$id/latest")
-      .response(asJson[Either[CdpApiError[CogniteId], Data[Items[StringDataPoint]]]])
+      .post(uri"$baseUri/latest")
+      .body(Items(Seq(CogniteId(id))))
+      .response(asJson[Either[CdpApiError[CogniteId], Items[StringDataPointsByIdResponse]]])
       .mapResponse {
         case Left(value) => throw value.error
         case Right(Left(cdpApiError)) => throw cdpApiError.asException(baseUri)
-        case Right(Right(value)) => extractor.extract(value).items.headOption
+        case Right(Right(value)) =>
+          extractor.extract(value).items.headOption.flatMap(_.datapoints.headOption)
       }
       .send()
+}
+
+object DataPointsResourceV1 {
+  // do something like add support for insert multiple?
 }
