+++ src/main/scala/com/cognite/sdk/scala/v1/resources/dataTypes.scala
@@ -1,3 +1,3 @@
-package com.cognite.sdk.scala.v1
+package com.cognite.sdk.scala.v1.resources
 
 final case class CogniteExternalId(id: String)
