+++ src/main/scala/com/cognite/sdk/scala/v1/resources/raw.scala
@@ -1,21 +1,11 @@
-package com.cognite.sdk.scala.v1
+package com.cognite.sdk.scala.v1.resources
 
-import com.cognite.sdk.scala.common.{
-  Auth,
-  CdpApiError,
-  CogniteId,
-  EitherDecoder,
-  Extractor,
-  ExtractorInstances,
-  Items,
-  ReadWritableResourceWithRetrieve,
-  WithId
-}
+import com.cognite.sdk.scala.common._
 import com.softwaremill.sttp._
 import com.softwaremill.sttp.circe._
-import io.circe.{Decoder, Encoder, Json}
 import io.circe.generic.auto._
 import io.circe.generic.semiauto.deriveEncoder
+import io.circe.{Decoder, Encoder, Json}
 
 final case class RawDatabase(name: String) extends WithId[String] {
   override val id: String = this.name
@@ -50,7 +40,6 @@ abstract class RawResource[R: Decoder, W: Decoder: Encoder, F[_], InternalId: En
 class RawDatabases[F[_]](project: String)(implicit auth: Auth, sttpBackend: SttpBackend[F, _])
     extends RawResource[RawDatabase, RawDatabase, F, RawDatabase, String] {
   def toInternalId(id: String): RawDatabase = RawDatabase(id)
-  implicit val extractor: Extractor[Id] = ExtractorInstances.idExtractor
   implicit val idEncoder: Encoder[RawDatabase] = deriveEncoder
   override val baseUri = uri"https://api.cognitedata.com/api/v1/projects/$project/raw/dbs"
 }
@@ -64,7 +53,6 @@ class RawTables[F[_]](project: String, database: String)(
     sttpBackend: SttpBackend[F, _]
 ) extends RawResource[RawTable, RawTable, F, RawTable, String] {
   def toInternalId(id: String): RawTable = RawTable(id)
-  implicit val extractor: Extractor[Id] = ExtractorInstances.idExtractor
   implicit val idEncoder: Encoder[RawTable] = deriveEncoder
   override val baseUri =
     uri"https://api.cognitedata.com/api/v1/projects/$project/raw/dbs/$database/tables"
@@ -75,13 +63,14 @@ class RawRows[F[_]](project: String, database: String, table: String)(
     sttpBackend: SttpBackend[F, _]
 ) extends RawResource[RawRow, RawRow, F, RawRowKey, String] {
   def toInternalId(id: String): RawRowKey = RawRowKey(id)
-  implicit val extractor: Extractor[Id] = ExtractorInstances.idExtractor
   implicit val idEncoder: Encoder[RawRowKey] = deriveEncoder
   override val baseUri =
     uri"https://api.cognitedata.com/api/v1/projects/$project/raw/dbs/$database/tables/$table/rows"
 
   // raw does not return the created rows in the response, so we'll always return an empty sequence.
-  override def createItems(items: Items[RawRow]): F[Response[Seq[RawRow]]] =
+  override def createItems(
+      items: Items[RawRow]
+  )(implicit extractor: Extractor[Id]): F[Response[Seq[RawRow]]] =
     request
       .post(baseUri)
       .body(items)
