+++ src/main/scala/com/cognite/sdk/scala/v1/resources/timeSeries.scala
@@ -1,4 +1,4 @@
-package com.cognite.sdk.scala.v1
+package com.cognite.sdk.scala.v1.resources
 
 import com.cognite.sdk.scala.common.{Auth, WithId}
 import com.softwaremill.sttp._
