+++ src/test/scala/com/cognite/sdk/scala/v06/DataPointsTest.scala
@@ -3,6 +3,7 @@ package com.cognite.sdk.scala.v06
 import java.util.UUID
 
 import com.cognite.sdk.scala.common.{DataPointsResourceBehaviors, SdkTest}
+import com.cognite.sdk.scala.v06.resources.TimeSeries
 
 class DataPointsTest extends SdkTest with DataPointsResourceBehaviors[Long] {
   private val client = new GenericClient()(auth, sttpBackend)
