+++ src/test/scala/com/cognite/sdk/scala/v06/TimeSeriesTest.scala
@@ -1,6 +1,7 @@
 package com.cognite.sdk.scala.v06
 
 import com.cognite.sdk.scala.common.{ReadableResourceBehaviors, SdkTest, WritableResourceBehaviors}
+import com.cognite.sdk.scala.v06.resources.{CreateTimeSeries, TimeSeries}
 
 class TimeSeriesTest extends SdkTest with ReadableResourceBehaviors with WritableResourceBehaviors {
   private val client = new GenericClient()(auth, sttpBackend)
