+++ src/test/scala/com/cognite/sdk/scala/v1/AssetsTest.scala
@@ -1,6 +1,7 @@
 package com.cognite.sdk.scala.v1
 
 import com.cognite.sdk.scala.common.{ReadableResourceBehaviors, SdkTest, WritableResourceBehaviors}
+import com.cognite.sdk.scala.v1.resources.{Asset, CreateAsset}
 
 class AssetsTest extends SdkTest with ReadableResourceBehaviors with WritableResourceBehaviors {
   private val client = new GenericClient()(auth, sttpBackend)
