+++ src/test/scala/com/cognite/sdk/scala/v1/ThreeDTest.scala
@@ -1,6 +1,7 @@
 package com.cognite.sdk.scala.v1
 
 import com.cognite.sdk.scala.common.{ReadableResourceBehaviors, SdkTest, WritableResourceBehaviors}
+import com.cognite.sdk.scala.v1.resources.{CreateThreeDModel, CreateThreeDRevision, ThreeDModel, ThreeDRevision}
 import io.circe.generic.auto._
 
 class ThreeDTest extends SdkTest with ReadableResourceBehaviors with WritableResourceBehaviors {
